+++ src/Pixes/pix_record.cpp
@@ -410,7 +410,7 @@ void pix_record :: codecMess(t_atom *argv)
   } else if (A_FLOAT==argv->a_type) {
     int id=atom_getint(argv);
     std::vector<std::string>codecs=m_handle->getCodecs();
-    if(id>0 && id<codecs.size())
+    if(id>0 && ((unsigned int)id)<codecs.size())
       sid=codecs[id];
     else {
       error("invalid codec# %d (0..%d)", id, codecs.size());
